Use the new API-endpoint.
Rather than reading Redis-data directly instead make a HTTP-request
to the `/global-stats` end-point of our API.  This ensures that we
don't have any special knowledge.

Initially I tried to use a JSONP request in the API, such that we
could let the javascript front-end make the request directly, but
because the API doesn't use SSL this failed with a mixed-security
warning.

It might be we shouldn't even show the damn spam-count, but since
we're doing so we'll pretend we're doing so cleanly.